Test Engine::joinStationAndMessageData method with a station and a message.
Test compinations are:
- In the first argument has a station data and the second has a message
  from an other station
- In the first argument has a message and the second has a station data
  which is not the same station where the message is from.
- In the first argument has a station data and the second has a message
  from the same station.

The tests doesn't test what the method does, actually. The method only
copies distance and bearing values from the first argument to the second
one in case if stationId values match and if the both argument have
bearing and/or distance columns. So, the tests are mostly useless.